+++ src/main/scala/cognite/spark/v1/DataModelInstancesHelper.scala
@@ -12,7 +12,7 @@ import org.apache.spark.sql.Row
 import org.apache.spark.sql.types.{DataType, DataTypes, StructType}
 
 // scalastyle:off cyclomatic.complexity
-object AlphaDataModelInstancesHelper {
+object DataModelInstancesHelper {
   private def unknownPropertyTypeMessage(a: Any) = s"Unknown property type $a."
 
   private def notValidPropertyTypeMessage(
