+++ src/main/scala/cognite/spark/v1/DataModelInstancesRelation.scala
@@ -2,7 +2,7 @@ package cognite.spark.v1
 
 import cats.effect.IO
 import cats.implicits._
-import cognite.spark.v1.AlphaDataModelInstancesHelper._
+import cognite.spark.v1.DataModelInstancesHelper._
 import com.cognite.sdk.scala.common.{
   CdpApiException,
   DSLAndFilter,
@@ -29,7 +29,7 @@ import org.apache.spark.sql.catalyst.expressions.GenericRow
 
 import scala.annotation.nowarn
 
-class AlphaDataModelInstanceRelation(
+class DataModelInstanceRelation(
     config: RelationConfig,
     spaceExternalId: String,
     modelExternalId: String,
