+++ src/test/scala/cognite/spark/v1/DataModelInstancesHelperTest.scala
@@ -1,11 +1,11 @@
 package cognite.spark.v1
 
-import cognite.spark.v1.AlphaDataModelInstancesHelper.getDirectRelationIdentifierProperty
+import cognite.spark.v1.DataModelInstancesHelper.getDirectRelationIdentifierProperty
 import com.cognite.sdk.scala.v1.DirectRelationIdentifier
 import org.apache.spark.sql.Row
 import org.scalatest.{FlatSpec, Matchers}
 
-class AlphaDataModelInstancesHelperTest extends FlatSpec with Matchers {
+class DataModelInstancesHelperTest extends FlatSpec with Matchers {
   it should "use first colon as the separator for spaceExternalId and externalId when get Direct Relation Identifier" in {
     val externalId = "instance-1"
     val row = Row.fromSeq(
